requirements.txt
torch
torchaudio
numpy
librosa
matplotlib
pyyaml
tqdm
scikit-learn